Add files via upload
There is a new trend in web design these days and that is to have a glowing button effect for call to action (CTA) buttons in an other-wise clean and flat design. In this post, I will create a simple button and apply glow effect with CSS on it when user hovers over it. We can also use the CSS presented below to add a permanent glow effect to CSS buttons.